fix: not needed entry point for variables removed, refactoring for process/task variables
