+++ activiti-cloud-services-query/activiti-cloud-services-query-rest/src/main/java/org/activiti/cloud/services/query/rest/assembler/TaskVariableResourceAssembler.java
@@ -21,7 +21,7 @@ import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
 
 import org.activiti.cloud.services.query.model.TaskVariableEntity;
 import org.activiti.cloud.services.query.resources.VariableResource;
-import org.activiti.cloud.services.query.rest.VariableController;
+import org.activiti.cloud.services.query.rest.TaskVariableController;
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.ResourceAssembler;
 import org.springframework.stereotype.Component;
@@ -31,7 +31,7 @@ public class TaskVariableResourceAssembler implements ResourceAssembler<TaskVari
 
     @Override
     public VariableResource toResource(TaskVariableEntity entity) {
-        Link selfRel = linkTo(methodOn(VariableController.class).findById(entity.getId())).withSelfRel();
+        Link selfRel = linkTo(methodOn(TaskVariableController.class).findById(entity.getId())).withSelfRel();
         return new VariableResource(entity, selfRel);
     }
