+++ activiti-cloud-services-query/activiti-cloud-services-query-rest/src/test/java/org/activiti/cloud/services/query/rest/ProcessInstanceEntityVariableEntityControllerIT.java
@@ -32,6 +32,7 @@ import java.util.Date;
 import java.util.UUID;
 
 import org.activiti.cloud.alfresco.argument.resolver.AlfrescoPageRequest;
+import org.activiti.cloud.services.query.app.repository.EntityFinder;
 import org.activiti.cloud.services.query.app.repository.VariableRepository;
 import org.activiti.cloud.services.query.model.ProcessVariableEntity;
 import org.junit.Test;
@@ -65,6 +66,9 @@ public class ProcessInstanceEntityVariableEntityControllerIT {
 
     @MockBean
     private VariableRepository variableRepository;
+    
+    @MockBean
+    private EntityFinder entityFinder;
 
     @Test
     public void getVariablesShouldReturnAllResultsUsingAlfrescoMetadataWhenMediaTypeIsApplicationJson() throws Exception {
